+++ src/main/kotlin/com/capitole/capitoleproductcatalogapi/application/getproductcatalog/dto/CategoryDTO.kt
@@ -1,4 +1,4 @@
-package com.capitole.capitoleproductcatalogapi.application.getproductcatalog
+package com.capitole.capitoleproductcatalogapi.application.getproductcatalog.dto
 
 import com.capitole.capitoleproductcatalogapi.domain.product.Category
