+++ src/main/kotlin/com/capitole/capitoleproductcatalogapi/infrastructure/configuration/ApplicationConfiguration.kt
@@ -1,6 +1,6 @@
 package com.capitole.capitoleproductcatalogapi.infrastructure.configuration
 
-import com.capitole.capitoleproductcatalogapi.application.getproductcatalog.GetProductCatalog
+import com.capitole.capitoleproductcatalogapi.application.getproductcatalog.dto.GetProductCatalog
 import com.capitole.capitoleproductcatalogapi.domain.discount.DiscountService
 import com.capitole.capitoleproductcatalogapi.domain.product.ProductRepository
 import org.springframework.context.annotation.Bean
