+++ src/test/kotlin/com/capitole/capitoleproductcatalogapi/infrastructure/GetProductCatalogControllerTest.kt
@@ -1,10 +1,10 @@
 package com.capitole.capitoleproductcatalogapi.infrastructure
 
-import com.capitole.capitoleproductcatalogapi.application.getproductcatalog.DiscountPercentageDTO
-import com.capitole.capitoleproductcatalogapi.application.getproductcatalog.GetProductCatalog
 import com.capitole.capitoleproductcatalogapi.application.getproductcatalog.PriceDTO
-import com.capitole.capitoleproductcatalogapi.application.getproductcatalog.ProductCatalogDTO
-import com.capitole.capitoleproductcatalogapi.application.getproductcatalog.toDTO
+import com.capitole.capitoleproductcatalogapi.application.getproductcatalog.dto.DiscountPercentageDTO
+import com.capitole.capitoleproductcatalogapi.application.getproductcatalog.dto.GetProductCatalog
+import com.capitole.capitoleproductcatalogapi.application.getproductcatalog.dto.ProductCatalogDTO
+import com.capitole.capitoleproductcatalogapi.application.getproductcatalog.dto.toDTO
 import com.capitole.capitoleproductcatalogapi.domain.pagination.PageRequest
 import com.capitole.capitoleproductcatalogapi.domain.product.Category
 import com.capitole.capitoleproductcatalogapi.domain.product.sort.SortField
